Rewrote code (details in description)
Changed CharacterController2D script to Player.

Renamed variables in Player script for consistency.

Removed unused code for crouch detection.

Rewrote all of the scene warping code to allow for A.I interaction in the future.

Warping now assigns a currentScene variable to the Player instance for access from the Enemy AI script or others.